Fix system stats print

+++ src/include/protocol_sensors.h
@@ -70,7 +70,7 @@ struct SystemStatus {
     uint16_t log_cnt;       // Number of total logged data sets since reset
     uint16_t rtc_s;         // Time since reset in seconds
     uint16_t log_ms;        // Time for writing the last log set to storage
-    uint32_t lock_wait_us;  // Delay until data structure could be locked
+    uint16_t lock_wait_us;  // Delay until data structure could be locked
 }__attribute__((packed));
 
 struct SensorData {
